feat: Claude Runtime Status Module + Preset Status Parity
Extract true status-string builders into claude-runtime-status.ts and add
readClaudeHookPresetStatus() for Codex/Claude adapter parity.

Changes:
- Create src/adapters/claude-runtime-status.ts with:
  - CLAUDE_ADDITIONAL_CONTEXT_MAX_CHARS constant
  - clampAdditionalContext() - 9000-char cap utility
  - sessionStartContext() - SessionStart context string
  - boundedFallbackContext() - fallback guidance string

- Refactor src/adapters/claude-runtime-hook.ts:
  - Import status utilities from new module
  - Re-export CLAUDE_ADDITIONAL_CONTEXT_MAX_CHARS for test compatibility
  - Keep buildPayloadContext() in place (payload assembly)

- Add readClaudeHookPresetStatus() to src/adapters/claude-hook-preset.ts:
  - Returns ClaudeHookPresetStatus with installedEvents, missingEvents,
    unexpectedFooksEvents, disabledByLocalSettings
  - Graceful error handling with valid:false + blocker on parse failure

- Refactor src/adapters/claude-status.ts:
  - hookStatus() delegates to readClaudeHookPresetStatus()
  - Keep manifest/adapter JSON checks as enrichment layer
  - Remove ~60 lines of duplicated preset-reading logic

Zero behavior change - pure refactoring with honest module boundaries.
